+++ examples/high_level_aux_plots.py
@@ -37,4 +37,4 @@ aux_results.atlas_second_tag = (
 
 # vertexing performance for b-jets
 logger.info("Plotting vertexing performance.")
-aux_results.plot_var_vtx_perf(flavour="bjets")
+aux_results.plot_var_vtx_perf(vtx_flavours=["bjets"], no_vtx_flavours=["ujets"])