+++ puma/var_vs_var.py
@@ -157,9 +157,6 @@ class VarVsVarPlot(PlotBase):
             raise ValueError("Not more than one ratio panel supported.")
         self.initialise_figure()
 
-    def __nonzero__(self):  # noqa: PLW3201
-        return bool(self.plot_objects)
-
     def add(self, curve: VarVsVar, key: str | None = None, reference: bool = False):
         """Adding VarVsVar object to figure.